+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- LSP create-note quick fixes now use wiki aliases and markdown link text as the new note's primary alias and heading.
+
 ## v0.4.0 - 2026-06-05
 
 ### Added